Use crossterm & ditch mio

+++ espmonitor/src/main.rs
@@ -45,7 +45,7 @@ fn parse_args() -> Result<Option<AppArgs>, Box<dyn Error>> {
         Ok(Some(AppArgs {
             chip,
             framework: Framework::default(),
-            speed: args.opt_value_from_fn("--speed", |s| s.parse::<u32>())?,
+            speed: args.opt_value_from_fn("--speed", |s| s.parse::<usize>())?,
             reset: args.contains("--reset") || !args.contains("--no-reset"),
             bin: args.opt_value_from_str("--bin")?,
             serial: args.free_from_str()?,
